+++ activestorage/test/service/s3_public_service_test.rb
@@ -42,7 +42,12 @@ if SERVICE_CONFIGURATIONS[:s3_public]
       request = Net::HTTP::Put.new uri.request_uri
       request.body = data
       request.add_field "Content-Type", "text/plain"
-      request.add_field "Content-MD5", checksum.digest
+      if checksum.digest == :MD5
+        request.add_field "Content-MD5", checksum.digest
+      else
+        request.add_field "x-amz-checksum-#{checksum.algorithm.downcase}", checksum.digest
+      end
+
       Net::HTTP.start(uri.host, uri.port, use_ssl: true) do |http|
         http.request request
       end